requirements.txt
telethon
python-dotenv
aiogram
python-dateutil
aiohttp